+++ CHANGELOG.md
@@ -26,6 +26,7 @@ All notable changes to this package will be documented in this file.
 - [TODO DOCS] Added: MyComponentUtility.MoveComponentInspectorToTop/ToBottom(Component component) utility methods
 - [TODO DOCS] Added: Contextual Menu Items to "Move Component To Top/To Bottom" in inspector
 - [TODO DOCS] Added: TopmostComponentAttribute to move on top in inspector automatically when it added
+- [TODO DOCS] Added: DisplayPreviewAttribute to display the property image of the specified size in the inspector
 
 ## [1.8.0] - 2023-10-09
 - Now it is possible to toggle some features that may lead to performance issues in editor in MyBox Window